# src/doculyzer/openapi.yaml
openapi: 3.0.3
info:
  title: Document Search API
  description: |
    A REST API for semantic document search that provides both basic and advanced search capabilities.
    The API supports embedding-based similarity search with optional content resolution, relationship information,
    and topic-based filtering for enhanced search precision.
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
paths:
...
  /api/search:
    post:
      tags:
        - Search
      summary: Basic search
      description: Search for elements and return basic results with similarity scores. Supports topic-based filtering for enhanced precision.
      operationId: search
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SearchRequest'
      responses:
        '200':
          description: Search completed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResponse'
        '400':
          description: Bad request - invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized - invalid or missing API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/search/advanced:
    post:
      tags:
        - Search
      summary: Advanced search
      description: Advanced search with full results including relationships and resolved content. Supports topic-based filtering for enhanced precision.
      operationId: advancedSearch
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdvancedSearchRequest'
      responses:
        '200':
          description: Advanced search completed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdvancedSearchResponse'
        '400':
          description: Bad request - invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized - invalid or missing API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/search/sources:
    post:
      tags:
        - Search
      summary: Get document sources
      description: Get unique document sources from search results. Supports topic-based filtering.
      operationId: getDocumentSources
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SourcesRequest'
      responses:
        '200':
          description: Document sources retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SourcesResponse'
        '400':
          description: Bad request - invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized - invalid or missing API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication
  schemas:
...
    SearchRequest:
      type: object
      required:
        - query
      properties:
        query:
          type: string
          description: The search query text
          example: "machine learning algorithms"
        limit:
          type: integer
          minimum: 1
          maximum: 100
          default: 10
          description: Maximum number of results to return
        min_score:
          type: number
          format: float
          minimum: 0.0
          maximum: 1.0
          default: 0.0
          description: Minimum similarity score threshold
        filter_criteria:
          type: object
          additionalProperties: true
          description: Optional filtering criteria
        include_topics:
          type: array
          items:
            type: string
          description: Topic LIKE patterns to include (e.g., ['security%', '%.policy%'])
          example: ["security%", "%.policy%"]
        exclude_topics:
          type: array
          items:
            type: string
          description: Topic LIKE patterns to exclude (e.g., ['deprecated%'])
          example: ["deprecated%", "draft%"]
        min_confidence:
          type: number
          format: float
          minimum: 0.0
          maximum: 1.0
          description: Minimum confidence threshold for topic results
          example: 0.7
        text:
          type: boolean
          default: false
          description: Whether to include resolved text content
        content:
          type: boolean
          default: false
          description: Whether to include resolved original content
        flat:
          type: boolean
          default: false
          description: Whether to return flat results instead of hierarchical
        include_parents:
          type: boolean
          default: true
          description: Whether to include parent elements in flat results
    AdvancedSearchRequest:
      type: object
      required:
        - query
      properties:
        query:
          type: string
          description: The search query text
          example: "deep learning frameworks"
        limit:
          type: integer
          minimum: 1
          maximum: 100
          default: 10
          description: Maximum number of results to return
        min_score:
          type: number
          format: float
          minimum: 0.0
          maximum: 1.0
          default: 0.0
          description: Minimum similarity score threshold
        filter_criteria:
          type: object
          additionalProperties: true
          description: Optional filtering criteria
        include_topics:
          type: array
          items:
            type: string
          description: Topic LIKE patterns to include (e.g., ['security%', '%.policy%'])
          example: ["security%", "%.policy%"]
        exclude_topics:
          type: array
          items:
            type: string
          description: Topic LIKE patterns to exclude (e.g., ['deprecated%'])
          example: ["deprecated%", "draft%"]
        min_confidence:
          type: number
          format: float
          minimum: 0.0
          maximum: 1.0
          description: Minimum confidence threshold for topic results
          example: 0.7
        resolve_content:
          type: boolean
          default: true
          description: Whether to resolve the original content
        include_relationships:
          type: boolean
          default: true
          description: Whether to include outgoing relationships
    SourcesRequest:
      type: object
      required:
        - query
      properties:
        query:
          type: string
          description: The search query text
          example: "artificial intelligence research"
        limit:
          type: integer
          minimum: 1
          maximum: 100
          default: 10
          description: Maximum number of results to consider
        min_score:
          type: number
          format: float
          minimum: 0.0
          maximum: 1.0
          default: 0.0
          description: Minimum similarity score threshold
        filter_criteria:
          type: object
          additionalProperties: true
          description: Optional filtering criteria
        include_topics:
          type: array
          items:
            type: string
          description: Topic LIKE patterns to include (e.g., ['security%', '%.policy%'])
          example: ["security%", "%.policy%"]
        exclude_topics:
          type: array
          items:
            type: string
          description: Topic LIKE patterns to exclude (e.g., ['deprecated%'])
          example: ["deprecated%", "draft%"]
        min_confidence:
          type: number
          format: float
          minimum: 0.0
          maximum: 1.0
          description: Minimum confidence threshold for topic results
          example: 0.7

    # Updated ElementBase schema to reflect new structure
...
    ElementHierarchical:
      allOf:
        - $ref: '#/components/schemas/ElementBase'
        - type: object
          properties:
            child_elements:
              type: array
              items:
                $ref: '#/components/schemas/ElementHierarchical'
              description: Child elements in the hierarchy
    ElementFlat:
      allOf:
        - $ref: '#/components/schemas/ElementBase'
        - type: object
          properties:
            path:
              type: string
              description: Path of element IDs from document ID to current element
...
    SearchResponse:
      type: object
      properties:
        results:
          type: array
          items:
            $ref: '#/components/schemas/SearchResult'
          description: List of search results
        total_results:
          type: integer
          description: Total number of results found
        query:
          type: string
          nullable: true
          description: The search query used
        filter_criteria:
          type: object
          additionalProperties: true
          nullable: true
          description: Filter criteria used
        # Topic filtering criteria
        include_topics:
          type: array
          items:
            type: string
          nullable: true
          description: Topic patterns that were included
        exclude_topics:
          type: array
          items:
            type: string
          nullable: true
          description: Topic patterns that were excluded
        min_confidence:
          type: number
          format: float
          nullable: true
          description: Minimum confidence threshold used
        search_type:
          type: string
          default: "embedding"
          description: Type of search performed (embedding, text, content, topic)
          enum: ["embedding", "text", "content", "topic"]
        min_score:
          type: number
          format: float
          default: 0.0
          description: Minimum score threshold used
        documents:
          type: array
          items:
            type: string
          description: Unique document sources
        search_tree:
          oneOf:
            - type: array
              items:
                $ref: '#/components/schemas/ElementHierarchical'
            - type: array
              items:
                $ref: '#/components/schemas/ElementFlat'
          nullable: true
          description: Tree structure or flat list of results
        content_resolved:
          type: boolean
          default: false
          description: Whether content was resolved during search
        text_resolved:
          type: boolean
          default: false
          description: Whether text was resolved during search
        # Topic-related metadata
        supports_topics:
          type: boolean
          default: false
          description: Whether the backend supports topics
        topic_statistics:
          type: object
          additionalProperties: true
          nullable: true
          description: Topic distribution statistics
...
    AdvancedSearchResponse:
      type: object
      properties:
        query:
          type: string
          description: The search query used
        total_results:
          type: integer
          description: Total number of results found
        min_score:
          type: number
          format: float
          description: Minimum score threshold used
        # Topic filtering criteria
        include_topics:
          type: array
          items:
            type: string
          nullable: true
          description: Topic patterns that were included
        exclude_topics:
          type: array
          items:
            type: string
          nullable: true
          description: Topic patterns that were excluded
        min_confidence:
          type: number
          format: float
          nullable: true
          description: Minimum confidence threshold used
        results:
          type: array
          items:
            $ref: '#/components/schemas/AdvancedSearchResult'
    SourcesResponse:
      type: object
      properties:
        query:
          type: string
          description: The search query used
        total_results:
          type: integer
          description: Total number of results considered
        document_sources:
          type: array
          items:
            type: string
          description: Unique document sources
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error type or code
        message:
          type: string
          description: Human-readable error message
      required:
        - error
        - message
...